gcc 2.95 compat
git-svn-id: file:///home/lennart/svn/public/ifplugd/trunk@116 2bf48fe7-cfc1-0310-909f-d9042e1e0fef

+++ src/nlapi.c
@@ -86,7 +86,7 @@ int nlapi_work(int block) {
         for (; bytes > 0; p = NLMSG_NEXT(p, bytes)) {
             struct callback_info *c;
             
-            if (!NLMSG_OK(p, bytes) || bytes < sizeof(struct nlmsghdr) || bytes < p->nlmsg_len) {
+            if (!NLMSG_OK(p, bytes) || (size_t) bytes < sizeof(struct nlmsghdr) || (size_t) bytes < p->nlmsg_len) {
                 daemon_log(LOG_ERR, "NLAPI: Packet too small or truncated!\n");
                 return -1;
             }
